Verify user still exists in db before allowing further user actions
These changes will further protect routes that require user verification before an action can be taken.
It confirms that the user still exists in the database in case the token is still valid but user was deleted or user details was modified within the period through which the token is still valid.
It also reduces the number of db query that occurs for actions that require the user details or record instance from the db by attaching the user instance on the request object before furthering down to the middleware chain.